fix many (64-bit) format warnings and unsigned/signed compare warnings
Observations:
 - 32-bit platform (linux):
   * size_t: 32-bit unsigned
   * ssize_t: 32-bit signed (should always be the same size as size_t)
   * off_t: 64-bit signed (should be always signed)
      (it is 64 bit as we use LARGE_FILE options, without we couldn't handle files > 4GB)
 - 64-bit platform:
   * size_t: 64-bit unsigned
   * ssize_t: 64-bit signed
   * off_t: 64-bit signed

Fixes:
 - off_t will be casted to (intmax_t) and formatted with "%jd"
 - size_t gets formatted with "%zu"
 - ssize_t gets formatted witz "%zd"
 - off_t/size_t comparisons:
   This is no problem if off_t is bigger then size_t; the compiler will just
   case size_t to off_t which is no problem in that case.

   Now, on 64-bit they have the same size but differ in signedness; i cast
   size_t to off_t manually anyway and just hope, that no one is using such big
   numbers (on 64-bit this is a really big number).
   This was mostly in mmap code, and as we only mmap small chunks i don't think this
   is a problem.


git-svn-id: svn://svn.lighttpd.net/lighttpd/trunk@2131 152afb58-edef-0310-8abb-c4023f1b3aa9

+++ src/mod_chunked.c
@@ -202,7 +202,7 @@ URIHANDLER_FUNC(mod_chunked_response_header) {
 		con->response.content_length = chunkqueue_length(in);
 	}
 	if(con->response.content_length >= 0) {
-		if (p->conf.debug > 0) TRACE("response content length known, disabling chunked encoding.  len=%ju", con->response.content_length);
+		if (p->conf.debug > 0) TRACE("response content length known, disabling chunked encoding.  len=%jd", (intmax_t) con->response.content_length);
 		use_chunked = 0;
 	} else if (con->request.http_method != HTTP_METHOD_HEAD) {
 		/* a HEAD request never gets a chunk-encoding, but might stay with keep-alive
@@ -346,7 +346,7 @@ URIHANDLER_FUNC(mod_chunked_encode_response_content) {
 		out->bytes_in += 5;
 	}
 
-	if (hctx->debug > 1) TRACE("chunk encoded: in=%lld, out=%lld", in->bytes_out, out->bytes_in);
+	if (hctx->debug > 1) TRACE("chunk encoded: in=%jd, out=%jd", (intmax_t) in->bytes_out, (intmax_t) out->bytes_in);
 
 	chunkqueue_remove_finished_chunks(in);
